Add support for downloading only security updates to the `update` command. (#300)
This adds support for a `--security-only` flag to allow updates to be
run just for modules with security releases. This also adds a way to
report on the avaiable updates without actually running them, as well as
an alternate output format that supports just listing the modules names.
This mimics the behavior of `drush pm-updatestatus --format=list` which
is useful for feeding the output to another process that actually runs
the output, like an Ansible playbook.