Add an update-status command.
Fixes #350. This command is similar to the `update` command but only
lists the available updates. It mirrors the `drush pm-updatestatus`
command. It allows for outputing a flat list of module names to feed
into another process, like an ansible script. It also includes an option
to list only updates that are flagged as security releases.

Conflicts:
  commands/update.bee.inc